+++ src/plone/restapi/types/adapters.py
@@ -205,6 +205,23 @@ class ListJsonSchemaProvider(CollectionJsonSchemaProvider):
         return info
 
 
+@adapter(IList, Interface, Interface)
+@implementer(IJsonSchemaProvider)
+class RelatedItemsSchemaProvider(ListJsonSchemaProvider):
+
+    def get_items(self):
+        """Get items properties."""
+        value_type_adapter = getMultiAdapter(
+            (self.field.value_type, self.context, self.request),
+            IJsonSchemaProvider)
+
+        # Prevent rendering all choices.
+        should_render_choices = self.field.__name__ != 'relatedItems'
+        value_type_adapter.should_render_choices = should_render_choices
+
+        return value_type_adapter.get_schema()
+
+
 @adapter(ISet, Interface, Interface)
 @implementer(IJsonSchemaProvider)
 class SetJsonSchemaProvider(CollectionJsonSchemaProvider):
@@ -226,6 +243,10 @@ class TupleJsonSchemaProvider(SetJsonSchemaProvider):
 @implementer(IJsonSchemaProvider)
 class ChoiceJsonSchemaProvider(DefaultJsonSchemaProvider):
 
+    # optionally prevent rendering all choices in the vocab,
+    # ie relatedItems, which contains UUIDs for all content in the site.
+    should_render_choices = True
+
     def get_type(self):
         return 'string'
 
@@ -241,7 +262,7 @@ class ChoiceJsonSchemaProvider(DefaultJsonSchemaProvider):
         else:
             vocabulary = self.field.vocabulary
 
-        if hasattr(vocabulary, '__iter__'):
+        if hasattr(vocabulary, '__iter__') and self.should_render_choices:
             for term in vocabulary:
                 title = translate(term.title, context=self.request)
                 choices.append((term.token, title))